Allow zero ports in port aliase
Change the form validation to allow the port number to be zero
so the form does not complain when showing or editing "any"
aliases. We will refuse to make changes to the alias object
in the API, and we want that error message to come through
instead of the validator in the form complaining.

Change-Id: I306176526e853b98a2623ea0fde933e4ef95e778
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>